+++ insts/base/MIX/MIX.cpp
@@ -19,7 +19,7 @@ MIX::~MIX()
 	delete [] in;
 }
 
-int MIX::init(float p[], int n_args)
+int MIX::init(double p[], int n_args)
 {
 // p0 = outsk; p1 = insk; p2 = duration (-endtime); p3 = amp; p4-n = channel mix matrix
 // we're stashing the setline info in gen table 1
