+++ insts/base/WAVETABLE/WAVETABLE.cpp
@@ -23,7 +23,7 @@ WAVETABLE::WAVETABLE() : Instrument()
 	branch = 0;
 }
 
-int WAVETABLE::init(float p[], int n_args)
+int WAVETABLE::init(double p[], int n_args)
 {
 // p0 = start; p1 = dur; p2 = amplitude; p3 = frequency; p4 = stereo spread;
 // real-time control enabled for p3 and p4
